Use review API instead of individual comments.
Adopt the review API for sending feedback. This results in fewer API
calls to github and more importantly fewer notification emails for
users.

I've had to restructure the Command class tree to allow type checks in
a saner way. This change also has the side effect of combining all the
issue comments into a single review comment which is again less noise.