Add concept of hunks.
Diffs are made of hunks that are not easily subdivided further. When
running fixers we'll want to find the hunks that intersect between the
original changes and the fixer's changes. By having hunks as a concept
we can work at that level of granularity and look for hunks that have
any overlapping lines.